Move 32-bit word into primitives
