+++ library/Rattletrap/AttributeMapping.hs
@@ -1,6 +1,6 @@
 module Rattletrap.AttributeMapping where
 
-import Rattletrap.Word32
+import Rattletrap.Primitive.Word32
 
 import qualified Data.Binary as Binary
