+++ library/Rattletrap/AttributeValue/LoadoutOnline.hs
@@ -1,7 +1,7 @@
 module Rattletrap.AttributeValue.LoadoutOnline where
 
 import Rattletrap.Primitive.CompressedWord
-import Rattletrap.Word32
+import Rattletrap.Primitive.Word32
 import Rattletrap.Primitive.Word8
 
 import qualified Control.Monad as Monad
