+++ library/Rattletrap/AttributeValue/Pickup.hs
@@ -1,6 +1,6 @@
 module Rattletrap.AttributeValue.Pickup where
 
-import Rattletrap.Word32
+import Rattletrap.Primitive.Word32
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
