+++ library/Rattletrap/AttributeValue/PrivateMatchSettings.hs
@@ -1,7 +1,7 @@
 module Rattletrap.AttributeValue.PrivateMatchSettings where
 
 import Rattletrap.Primitive.Text
-import Rattletrap.Word32
+import Rattletrap.Primitive.Word32
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
