+++ library/Rattletrap/KeyFrame.hs
@@ -1,7 +1,7 @@
 module Rattletrap.KeyFrame where
 
 import Rattletrap.Primitive.Float32
-import Rattletrap.Word32
+import Rattletrap.Primitive.Word32
 
 import qualified Data.Binary as Binary
