+++ library/Rattletrap/ReplicationValue/Spawned.hs
@@ -5,7 +5,7 @@ import Rattletrap.ClassAttributeMap
 import Rattletrap.Primitive.CompressedWord
 import Rattletrap.Initialization
 import Rattletrap.Primitive.Text
-import Rattletrap.Word32
+import Rattletrap.Primitive.Word32
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
