+++ library/Rattletrap/Section.hs
@@ -1,7 +1,7 @@
 module Rattletrap.Section where
 
 import Rattletrap.Crc
-import Rattletrap.Word32
+import Rattletrap.Primitive.Word32
 
 import qualified Control.Monad as Monad
 import qualified Data.Binary as Binary
